+++ lib/commands/command_list.dart
@@ -5,12 +5,15 @@ import 'package:easy_init_cli/commands/run_build_runner/run_build_runner.dart';
 import 'package:easy_init_cli/commands/update/update.dart';
 import 'package:easy_init_cli/interfaces/command.dart';
 
+import 'create_helpers/create_helpers.dart';
+
 List<Command> commands = [
   CommandParent(
     "create",
     [
       CreateProject(),
       CreateFeature(),
+      CreateHelpers(),
     ],
   ),
   RunBuildRunner(),