+++ lib/core/structure/tdd_clean_structure/file_contents/number_trivia_content.dart
@@ -241,9 +241,9 @@ const triviaScreen = '''
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../../../../core/extensions/extensions.dart';
 import '../../../../common/widgets/loading.dart';
-import '../../../../common/widgets/space.dart'; // Importing custom widget for spacing
-import '../../../../core/theme/colors.dart'; // Importing custom theme colors
+import '../../../../core/theme/app_colors.dart'; // Importing custom theme colors
 import '../blocs/number_trivia_bloc/number_trivia_bloc.dart'; // Importing the NumberTriviaBloc
 
 /// to validate form [_formKey] is used
@@ -254,13 +254,9 @@ class NumberTriviaScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    /// The [theme] variable holds the current theme information.
-    /// This allows you to implement a UI that adapts to light and dark themes.
-    ThemeData theme = Theme.of(context); // Getting current theme data
-    Size size = MediaQuery.of(context).size; // Getting screen size
-    NumberTriviaBloc triviaBloc = context.read<
-        NumberTriviaBloc>(); // Getting instance of NumberTriviaBloc using context
-
+    /// Getting instance of NumberTriviaBloc using context
+    NumberTriviaBloc triviaBloc = context.read<NumberTriviaBloc>();
+    AppColors? appColors = context.appColors;
     return Scaffold(
       appBar: AppBar(
         title: const Text("Number Trivia by Easy Init"), // AppBar title
@@ -287,14 +283,16 @@ class NumberTriviaScreen extends StatelessWidget {
                     Expanded(
                       child: Center(
                         child: Container(
-                          width: size.width,
+                          width: context.screenWidth,
                           constraints: const BoxConstraints(minHeight: 40),
                           decoration: BoxDecoration(
                             borderRadius: BorderRadius.circular(10),
-                            color: theme.brightness == Brightness.dark
-                                ? const Color.fromARGB(255, 22, 22, 22)
-                                : const Color.fromARGB(255, 244, 244,
-                                    244), // Setting container background color
+                            color: context.setThemeBasedColor(
+                              darkThemeColor:
+                                  const Color.fromARGB(255, 22, 22, 22),
+                              lightThemeColor:
+                                  const Color.fromARGB(255, 244, 244, 244),
+                            ), // Setting container background color
                           ),
                           child: Padding(
                             padding: const EdgeInsets.all(8.0),
@@ -321,13 +319,14 @@ class NumberTriviaScreen extends StatelessWidget {
                         ),
                       ),
                     ),
-                    Space.y(10),
+                    10.verticalSpace,
                     TextFormField(
                       controller: triviaBloc
                           .numberController, // Binding the TextEditingController from bloc
                       keyboardType:
                           TextInputType.number, // Allowing only number input
                       decoration: InputDecoration(
+                        hintText: "Enter a number",
                         contentPadding: const EdgeInsetsDirectional.symmetric(
                           horizontal: 15,
                           vertical: 5,
@@ -343,13 +342,19 @@ class NumberTriviaScreen extends StatelessWidget {
                         return null;
                       },
                     ),
-                    Space.y(10), // Adding vertical space
+                    // Adding vertical space
+                    10.verticalSpace,
+
                     SizedBox(
-                      width: size.width,
+                      height: 44,
+                      width: context.screenWidth,
                       child: ElevatedButton(
                         style: ElevatedButton.styleFrom(
-                          backgroundColor: AppColors
-                              .blackColor, // Setting button background color
+                          // Setting button background color
+                          backgroundColor: context.setThemeBasedColor(
+                            darkThemeColor: appColors?.white,
+                            lightThemeColor: appColors?.black,
+                          ),
                         ),
                         onPressed: () {
                           if (_formKey.currentState!.validate()) {
@@ -359,14 +364,16 @@ class NumberTriviaScreen extends StatelessWidget {
                         },
                         child: Text(
                           "Get Trivia", // Button text
-                          style: theme.textTheme.labelLarge?.copyWith(
-                            color: AppColors
-                                .whiteColor, // Setting button text color
+                          style: context.labelLarge(
+                            color: context.setThemeBasedColor(
+                              darkThemeColor: appColors?.black,
+                              lightThemeColor: appColors?.white,
+                            ),
                           ),
                         ),
                       ),
                     ),
-                    Space.y(10), // Adding vertical space
+                    10.verticalSpace // Adding vertical space
                   ],
                 ),
               ),