+++ lib/core/structure/tdd_clean_structure/tdd_clean_structure.dart
@@ -29,6 +29,11 @@ class TddCleanStructure extends Structure {
             path: "lib/common/widgets",
           ),
         ),
+        CleanDirName.commonHelpers: Directory(
+          replaceAsExpected(
+            path: "lib/common/helpers",
+          ),
+        ),
         CleanDirName.core: Directory(
           replaceAsExpected(
             path: "lib/core",
@@ -49,6 +54,11 @@ class TddCleanStructure extends Structure {
             path: "lib/core/dependancy_injection",
           ),
         ),
+        CleanDirName.extensions: Directory(
+          replaceAsExpected(
+            path: "lib/core/extensions",
+          ),
+        ),
         CleanDirName.failures: Directory(
           replaceAsExpected(
             path: "lib/core/failures",
@@ -67,6 +77,11 @@ class TddCleanStructure extends Structure {
         CleanDirName.theme: Directory(
           replaceAsExpected(
             path: "lib/core/theme",
+          ),
+        ),
+        CleanDirName.typography: Directory(
+          replaceAsExpected(
+            path: "lib/core/theme/typography",
           ),
         ),
         CleanDirName.config: Directory(
@@ -173,11 +188,6 @@ class TddCleanStructure extends Structure {
         ),
         FileModel(
           directoryStructure[CleanDirName.commonWidgets]!.path,
-          "space.dart",
-          spaceContent,
-        ),
-        FileModel(
-          directoryStructure[CleanDirName.commonWidgets]!.path,
           "loading.dart",
           loadingContent,
         ),
@@ -185,6 +195,11 @@ class TddCleanStructure extends Structure {
           directoryStructure[CleanDirName.commonWidgets]!.path,
           "responsive.dart",
           responsiveContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.commonHelpers]!.path,
+          ".gitkeep",
+          '',
         ),
         FileModel(
           directoryStructure[CleanDirName.apiEndpoints]!.path,
@@ -205,6 +220,36 @@ class TddCleanStructure extends Structure {
           directoryStructure[CleanDirName.modules]!.path,
           "dio_module.dart",
           dioModuleContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "app_navigation_ext.dart",
+          navigationExtContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "date_ext.dart",
+          dateExtContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "string_ext.dart",
+          stringExtContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "theme_ext.dart",
+          themeExtContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "number_ext.dart",
+          numberExtContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.extensions]!.path,
+          "extensions.dart",
+          extensionsContent,
         ),
         FileModel(
           directoryStructure[CleanDirName.failures]!.path,
@@ -229,12 +274,32 @@ class TddCleanStructure extends Structure {
         FileModel(
           directoryStructure[CleanDirName.theme]!.path,
           "theme.dart",
-          themeContent,
+          theme,
         ),
         FileModel(
           directoryStructure[CleanDirName.theme]!.path,
-          "colors.dart",
+          "app_colors.dart",
           colorsContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.theme]!.path,
+          "app_theme.dart",
+          themeContent,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.typography]!.path,
+          "typography.dart",
+          typography,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.typography]!.path,
+          "app_font_weight.dart",
+          appFontWeight,
+        ),
+        FileModel(
+          directoryStructure[CleanDirName.typography]!.path,
+          "ui_text_style.dart",
+          textStyles,
         ),
         FileModel(
           directoryStructure[CleanDirName.features]!.path,
@@ -400,15 +465,18 @@ class CleanDirName {
   //configurations
   static String common = "common";
   static String commonWidgets = "widgets";
+  static String commonHelpers = "helpers";
   static String features = "features";
   static String core = "core";
   static String apiEndpoints = "apiEndpoints";
   static String baseUsecase = "base_usecase";
   static String dependancyInjection = "dependancy_injection";
+  static String extensions = "extensions";
   static String failures = "failures";
   static String network = "network";
   static String routes = "routes";
   static String theme = "theme";
+  static String typography = "typography";
   static String config = "config";
   static String modules = "modules";
   //features